centralize hosted diagnostic redaction
