+++ packages/assistant-runtime/src/hosted-runtime/system-mailbox.ts
@@ -2,6 +2,9 @@ import { rm } from "node:fs/promises";
 import path from "node:path";
 
 import {
+  readHostedRuntimeSafeErrorText,
+} from "@murphai/hosted-execution";
+import {
   type HostedExecutionSystemWake,
 } from "@murphai/hosted-execution/contracts";
 import type {
@@ -17,9 +20,6 @@ import {
 import {
   executeHostedMailboxEvent,
 } from "./events.ts";
-import {
-  readHostedRuntimeSafeErrorText,
-} from "./diagnostic-redaction.ts";
 import type {
   HostedMailboxItemImportOutcome,
   HostedMailboxResolvedImportItem,
